refactor(sbom): address PR review feedback for exploit intelligence
- Move state-mapping logic from queries/ to hooks/domain-controls/,
  creating useExploitIntelligenceOfSbom following the existing
  domain-controls pattern (useVulnerabilitiesOfSbom)
- Replace raw axios with shared API client from apiInit
- Remove custom CSS in favor of PatternFly defaults
- Remove requestAnalysisEligible prop (all severities are eligible)
- Improve test readability with makeMultiComponentJob helper and
  it.each for repetitive cases
