+++ client/src/app/hooks/domain-controls/useExploitIntelligenceOfSbom.test.ts
@@ -1,8 +1,6 @@
-import {
-  type ExploitIntelligenceJobSummary,
-  buildStateMap,
-  jobToState,
-} from "./exploit-intelligence";
+import type { ExploitIntelligenceJobSummary } from "@app/queries/exploit-intelligence";
+
+import { buildStateMap, jobToState } from "./useExploitIntelligenceOfSbom";
 
 const makeJob = (
   overrides: Partial<ExploitIntelligenceJobSummary> = {},
@@ -25,22 +23,31 @@ const makeJob = (
   ...overrides,
 });
 
-describe("jobToState", () => {
-  it("maps pending job to in_progress", () => {
-    const result = jobToState(makeJob({ status: "pending", finding: null }));
-    expect(result).toEqual({
-      kind: "finding",
-      finding: { variant: "in_progress" },
-    });
+const makeMultiComponentJob = (
+  overrides: Partial<ExploitIntelligenceJobSummary> = {},
+): ExploitIntelligenceJobSummary =>
+  makeJob({
+    product_id: "product-1",
+    total_components: 5,
+    completed_components: 5,
+    failed_components: 0,
+    vulnerable_components: 0,
+    not_vulnerable_components: 5,
+    uncertain_components: 0,
+    ...overrides,
   });
 
-  it("maps running job to in_progress", () => {
-    const result = jobToState(makeJob({ status: "running", finding: null }));
-    expect(result).toEqual({
-      kind: "finding",
-      finding: { variant: "in_progress" },
-    });
-  });
+describe("jobToState", () => {
+  it.each(["pending", "running"] as const)(
+    "maps %s job to in_progress",
+    (status) => {
+      const result = jobToState(makeJob({ status, finding: null }));
+      expect(result).toEqual({
+        kind: "finding",
+        finding: { variant: "in_progress" },
+      });
+    },
+  );
 
   it("maps failed job to failed", () => {
     const result = jobToState(makeJob({ status: "failed", finding: null }));
@@ -50,7 +57,31 @@ describe("jobToState", () => {
     });
   });
 
-  it("maps completed + vulnerable to vulnerable with reportUrl", () => {
+  it.each([
+    { finding: "vulnerable" as const, expectedVariant: "vulnerable" },
+    { finding: "not_vulnerable" as const, expectedVariant: "not_vulnerable" },
+    { finding: "uncertain" as const, expectedVariant: "uncertain" },
+  ])(
+    "maps completed + $finding to $expectedVariant",
+    ({ finding, expectedVariant }) => {
+      const result = jobToState(makeJob({ status: "completed", finding }));
+      expect(result.kind).toBe("finding");
+      if (result.kind === "finding") {
+        expect(result.finding.variant).toBe(expectedVariant);
+      }
+    },
+  );
+
+  it("maps completed with null finding to uncertain", () => {
+    const result = jobToState(makeJob({ status: "completed", finding: null }));
+    expect(result).toEqual({
+      kind: "finding",
+      finding: { variant: "uncertain" },
+      reportUrl: undefined,
+    });
+  });
+
+  it("includes reportUrl when present", () => {
     const result = jobToState(
       makeJob({
         status: "completed",
@@ -65,62 +96,11 @@ describe("jobToState", () => {
     });
   });
 
-  it("maps completed + not_vulnerable to not_vulnerable", () => {
+  it("includes count and breakdown for multi-component completed job", () => {
     const result = jobToState(
-      makeJob({ status: "completed", finding: "not_vulnerable" }),
-    );
-    expect(result).toEqual({
-      kind: "finding",
-      finding: { variant: "not_vulnerable" },
-      reportUrl: undefined,
-    });
-  });
-
-  it("maps completed + uncertain to uncertain", () => {
-    const result = jobToState(
-      makeJob({ status: "completed", finding: "uncertain" }),
-    );
-    expect(result).toEqual({
-      kind: "finding",
-      finding: { variant: "uncertain" },
-      reportUrl: undefined,
-    });
-  });
-
-  it("maps completed with null finding to uncertain", () => {
-    const result = jobToState(makeJob({ status: "completed", finding: null }));
-    expect(result).toEqual({
-      kind: "finding",
-      finding: { variant: "uncertain" },
-      reportUrl: undefined,
-    });
-  });
-
-  it("omits breakdown for multi-component running job (in_progress)", () => {
-    const result = jobToState(
-      makeJob({
-        status: "running",
-        finding: null,
-        product_id: "product-1",
-        total_components: 5,
-        completed_components: 3,
-        failed_components: 0,
-      }),
-    );
-    expect(result).toEqual({
-      kind: "finding",
-      finding: { variant: "in_progress" },
-    });
-  });
-
-  it("includes count and breakdown for multi-component completed vulnerable job", () => {
-    const result = jobToState(
-      makeJob({
-        status: "completed",
+      makeMultiComponentJob({
         finding: "vulnerable",
         report_url: "https://example.com/report",
-        product_id: "product-1",
-        total_components: 5,
         completed_components: 4,
         failed_components: 1,
         vulnerable_components: 2,
@@ -144,50 +124,33 @@ describe("jobToState", () => {
     });
   });
 
-  it("includes count and breakdown for multi-component completed uncertain job", () => {
+  it("uses the variant-specific count for uncertain findings", () => {
     const result = jobToState(
-      makeJob({
-        status: "completed",
+      makeMultiComponentJob({
         finding: "uncertain",
-        product_id: "product-1",
         total_components: 74,
         completed_components: 74,
-        failed_components: 0,
-        vulnerable_components: 0,
         not_vulnerable_components: 70,
         uncertain_components: 4,
       }),
     );
-    expect(result).toEqual({
-      kind: "finding",
-      finding: {
-        variant: "uncertain",
-        count: 4,
-        breakdown: {
-          vulnerableCount: 0,
-          uncertainCount: 4,
-          notVulnerableCount: 70,
-          failedCount: 0,
-        },
-      },
-      reportUrl: undefined,
-    });
+
+    if (result.kind === "finding") {
+      expect(result.finding.variant).toBe("uncertain");
+      expect((result.finding as { count?: number }).count).toBe(4);
+    }
   });
 
-  it("includes breakdown without count for multi-component not_vulnerable job", () => {
+  it("includes breakdown without count for not_vulnerable", () => {
     const result = jobToState(
-      makeJob({
-        status: "completed",
+      makeMultiComponentJob({
         finding: "not_vulnerable",
-        product_id: "product-1",
         total_components: 88,
         completed_components: 88,
-        failed_components: 0,
-        vulnerable_components: 0,
         not_vulnerable_components: 88,
-        uncertain_components: 0,
       }),
     );
+
     expect(result).toEqual({
       kind: "finding",
       finding: {
@@ -205,10 +168,8 @@ describe("jobToState", () => {
 
   it("defaults component counts to 0 when null", () => {
     const result = jobToState(
-      makeJob({
-        status: "completed",
+      makeMultiComponentJob({
         finding: "vulnerable",
-        product_id: "product-1",
         total_components: 3,
         completed_components: null,
         failed_components: null,
@@ -217,20 +178,15 @@ describe("jobToState", () => {
         uncertain_components: null,
       }),
     );
-    expect(result).toEqual({
-      kind: "finding",
-      finding: {
-        variant: "vulnerable",
-        count: 0,
-        breakdown: {
-          vulnerableCount: 0,
-          uncertainCount: 0,
-          notVulnerableCount: 0,
-          failedCount: 0,
-        },
-      },
-      reportUrl: undefined,
-    });
+
+    if (result.kind === "finding" && "breakdown" in result.finding) {
+      expect(result.finding.breakdown).toEqual({
+        vulnerableCount: 0,
+        uncertainCount: 0,
+        notVulnerableCount: 0,
+        failedCount: 0,
+      });
+    }
   });
 
   it("omits breakdown when product_id is null", () => {
@@ -249,13 +205,21 @@ describe("jobToState", () => {
     });
   });
 
+  it("omits breakdown for in-progress multi-component job", () => {
+    const result = jobToState(
+      makeMultiComponentJob({ status: "running", finding: null }),
+    );
+    expect(result).toEqual({
+      kind: "finding",
+      finding: { variant: "in_progress" },
+    });
+  });
+
   it("omits breakdown for failed multi-component job", () => {
     const result = jobToState(
-      makeJob({
+      makeMultiComponentJob({
         status: "failed",
         finding: null,
-        product_id: "product-1",
-        total_components: 5,
         completed_components: 2,
         failed_components: 3,
       }),
@@ -273,62 +237,55 @@ describe("buildStateMap", () => {
   });
 
   it("builds map keyed by vulnerability_id", () => {
-    const jobs = [
+    const map = buildStateMap([
       makeJob({ vulnerability_id: "CVE-2024-0001", finding: "vulnerable" }),
       makeJob({
         vulnerability_id: "CVE-2024-0002",
         finding: "not_vulnerable",
         id: "job-2",
       }),
-    ];
-    const map = buildStateMap(jobs);
+    ]);
+
     expect(Object.keys(map)).toHaveLength(2);
     expect(map["CVE-2024-0001"]?.kind).toBe("finding");
     expect(map["CVE-2024-0002"]?.kind).toBe("finding");
   });
 
   it("prefers in-progress job over completed for same vulnerability", () => {
-    const jobs = [
+    const map = buildStateMap([
       makeJob({
         id: "job-1",
-        vulnerability_id: "CVE-2024-0001",
         status: "completed",
         finding: "vulnerable",
       }),
       makeJob({
         id: "job-2",
-        vulnerability_id: "CVE-2024-0001",
         status: "running",
         finding: null,
       }),
-    ];
-    const map = buildStateMap(jobs);
+    ]);
+
     expect(map["CVE-2024-0001"]).toEqual({
       kind: "finding",
       finding: { variant: "in_progress" },
     });
   });
 
-  it("preserves breakdown in state map for multi-component jobs", () => {
-    const jobs = [
-      makeJob({
-        vulnerability_id: "CVE-2024-0001",
-        status: "completed",
+  it("preserves breakdown for multi-component jobs", () => {
+    const map = buildStateMap([
+      makeMultiComponentJob({
         finding: "vulnerable",
-        product_id: "product-1",
-        total_components: 5,
         completed_components: 4,
         failed_components: 1,
         vulnerable_components: 2,
         not_vulnerable_components: 1,
         uncertain_components: 1,
       }),
-    ];
-    const map = buildStateMap(jobs);
-    const state = map["CVE-2024-0001"];
-    expect(state?.kind).toBe("finding");
-    if (state?.kind === "finding") {
-      expect(state.finding).toEqual({
+    ]);
+
+    expect(map["CVE-2024-0001"]).toEqual({
+      kind: "finding",
+      finding: {
         variant: "vulnerable",
         count: 2,
         breakdown: {
@@ -337,7 +294,8 @@ describe("buildStateMap", () => {
           notVulnerableCount: 1,
           failedCount: 1,
         },
-      });
-    }
+      },
+      reportUrl: undefined,
+    });
   });
 });
