+++ client/src/app/hooks/domain-controls/useExploitIntelligenceOfSbom.ts
@@ -1,0 +1,118 @@
+import React from "react";
+
+import type {
+  ExploitIntelligenceCellState,
+  ExploitIntelligenceFinding,
+  ExploitIntelligenceFindingBreakdown,
+} from "@app/pages/sbom-details/components/exploit-intelligence-analysis-cell";
+import {
+  type ExploitIntelligenceJobSummary,
+  useFetchExploitIntelligenceJobs,
+} from "@app/queries/exploit-intelligence";
+
+type Finding = "vulnerable" | "not_vulnerable" | "uncertain";
+
+const buildBreakdown = (
+  job: ExploitIntelligenceJobSummary,
+): ExploitIntelligenceFindingBreakdown | undefined => {
+  if (job.product_id == null || job.total_components == null) {
+    return undefined;
+  }
+  return {
+    vulnerableCount: job.vulnerable_components ?? 0,
+    uncertainCount: job.uncertain_components ?? 0,
+    notVulnerableCount: job.not_vulnerable_components ?? 0,
+    failedCount: job.failed_components ?? 0,
+  };
+};
+
+const buildCompletedFinding = (
+  variant: Finding,
+  breakdown: ExploitIntelligenceFindingBreakdown | undefined,
+): ExploitIntelligenceFinding => {
+  switch (variant) {
+    case "vulnerable":
+      return {
+        variant: "vulnerable",
+        count: breakdown ? breakdown.vulnerableCount : undefined,
+        breakdown,
+      };
+    case "uncertain":
+      return {
+        variant: "uncertain",
+        count: breakdown ? breakdown.uncertainCount : undefined,
+        breakdown,
+      };
+    case "not_vulnerable":
+      return { variant: "not_vulnerable", breakdown };
+  }
+};
+
+export const jobToState = (
+  job: ExploitIntelligenceJobSummary,
+): ExploitIntelligenceCellState => {
+  if (job.status === "pending" || job.status === "running") {
+    return {
+      kind: "finding",
+      finding: { variant: "in_progress" },
+    };
+  }
+
+  if (job.status === "failed") {
+    return {
+      kind: "finding",
+      finding: { variant: "failed" },
+    };
+  }
+
+  const variant = job.finding ?? "uncertain";
+  const breakdown = buildBreakdown(job);
+  return {
+    kind: "finding",
+    finding: buildCompletedFinding(variant, breakdown),
+    reportUrl: job.report_url ?? undefined,
+  };
+};
+
+const shouldReplace = (
+  current: ExploitIntelligenceCellState,
+  candidate: ExploitIntelligenceJobSummary,
+): boolean => {
+  if (current.kind === "not_run") return true;
+  if (candidate.status === "pending" || candidate.status === "running")
+    return true;
+  if (
+    candidate.status === "completed" &&
+    current.finding.variant !== "vulnerable" &&
+    current.finding.variant !== "not_vulnerable" &&
+    current.finding.variant !== "uncertain"
+  )
+    return true;
+  return false;
+};
+
+export const buildStateMap = (
+  jobs: ExploitIntelligenceJobSummary[],
+): Record<string, ExploitIntelligenceCellState> => {
+  const map: Record<string, ExploitIntelligenceCellState> = {};
+  for (const job of jobs) {
+    const existing = map[job.vulnerability_id];
+    if (!existing || shouldReplace(existing, job)) {
+      map[job.vulnerability_id] = jobToState(job);
+    }
+  }
+  return map;
+};
+
+export const useExploitIntelligenceOfSbom = (sbomId: string) => {
+  const { jobs, isFetching, fetchError } =
+    useFetchExploitIntelligenceJobs(sbomId);
+
+  const stateMap = React.useMemo(() => buildStateMap(jobs), [jobs]);
+
+  return {
+    stateMap,
+    isFetching,
+    fetchError,
+  };
+};
