+++ client/src/app/pages/sbom-details/components/exploit-intelligence-analysis-cell.test.tsx
@@ -14,7 +14,6 @@ describe("ExploitIntelligenceAnalysisCell", () => {
         vulnerabilityIdentifier={CVE_ID}
         state={{ kind: "not_run" }}
         onRequestAnalysis={vi.fn()}
-        requestAnalysisEligible={true}
       />,
     );
 
@@ -23,23 +22,6 @@ describe("ExploitIntelligenceAnalysisCell", () => {
     });
     expect(button).toBeInTheDocument();
     expect(button).toBeEnabled();
-  });
-
-  it("renders disabled button when requestAnalysisEligible is false", () => {
-    render(
-      <ExploitIntelligenceAnalysisCell
-        vulnerabilityIdentifier={CVE_ID}
-        state={{ kind: "not_run" }}
-        onRequestAnalysis={vi.fn()}
-        requestAnalysisEligible={false}
-      />,
-    );
-
-    const button = screen.getByRole("button", {
-      name: /request.*analysis/i,
-    });
-    expect(button).toBeInTheDocument();
-    expect(button).toBeDisabled();
   });
 
   it.each<{
@@ -72,7 +54,6 @@ describe("ExploitIntelligenceAnalysisCell", () => {
         vulnerabilityIdentifier={CVE_ID}
         state={state}
         onRequestAnalysis={vi.fn()}
-        requestAnalysisEligible={true}
       />,
     );
 
@@ -100,7 +81,6 @@ describe("ExploitIntelligenceAnalysisCell", () => {
         vulnerabilityIdentifier={CVE_ID}
         state={state}
         onRequestAnalysis={vi.fn()}
-        requestAnalysisEligible={true}
       />,
     );
 
@@ -127,7 +107,6 @@ describe("ExploitIntelligenceAnalysisCell", () => {
         vulnerabilityIdentifier={CVE_ID}
         state={state}
         onRequestAnalysis={vi.fn()}
-        requestAnalysisEligible={true}
       />,
     );
 
@@ -146,7 +125,6 @@ describe("ExploitIntelligenceAnalysisCell", () => {
         vulnerabilityIdentifier={CVE_ID}
         state={state}
         onRequestAnalysis={vi.fn()}
-        requestAnalysisEligible={true}
       />,
     );
 
@@ -167,7 +145,6 @@ describe("ExploitIntelligenceAnalysisCell", () => {
         vulnerabilityIdentifier={CVE_ID}
         state={{ kind: "not_run" }}
         onRequestAnalysis={onRequestAnalysis}
-        requestAnalysisEligible={true}
       />,
     );
 
@@ -191,7 +168,6 @@ describe("ExploitIntelligenceAnalysisCell", () => {
         vulnerabilityIdentifier={CVE_ID}
         state={state}
         onRequestAnalysis={vi.fn()}
-        requestAnalysisEligible={true}
       />,
     );
 
@@ -209,7 +185,6 @@ describe("ExploitIntelligenceAnalysisCell", () => {
         vulnerabilityIdentifier={CVE_ID}
         state={state}
         onRequestAnalysis={vi.fn()}
-        requestAnalysisEligible={true}
       />,
     );
