+++ client/src/app/pages/sbom-details/components/exploit-intelligence-analysis-cell.tsx
@@ -3,8 +3,6 @@ import type React from "react";
 import { Button, Flex, Label, Spinner, Tooltip } from "@patternfly/react-core";
 import ExclamationCircleIcon from "@patternfly/react-icons/dist/esm/icons/exclamation-circle-icon";
 
-import "./exploit-intelligence.css";
-
 export interface ExploitIntelligenceFindingBreakdown {
   vulnerableCount: number;
   uncertainCount: number;
@@ -61,11 +59,7 @@ export interface ExploitIntelligenceAnalysisCellProps {
   vulnerabilityIdentifier: string;
   state: ExploitIntelligenceCellState;
   onRequestAnalysis: (vulnerabilityIdentifier: string) => void;
-  requestAnalysisEligible?: boolean;
 }
-
-const requestIneligibleTooltip =
-  "Exploit Intelligence analysis can be requested only for Critical and High severity vulnerabilities.";
 
 const findingVariantsWithReport: ExploitIntelligenceFinding["variant"][] = [
   "vulnerable",
@@ -149,7 +143,6 @@ const FindingLabel: React.FC<{ finding: ExploitIntelligenceFinding }> = ({
         return (
           <Label
             color="grey"
-            className="exploit-intelligence-in-progress-label"
             icon={
               <Spinner
                 isInline
@@ -214,39 +207,19 @@ const FindingWithOptionalReportLink: React.FC<{
 
 export const ExploitIntelligenceAnalysisCell: React.FC<
   ExploitIntelligenceAnalysisCellProps
-> = ({
-  vulnerabilityIdentifier,
-  state,
-  onRequestAnalysis,
-  requestAnalysisEligible = true,
-}) => {
+> = ({ vulnerabilityIdentifier, state, onRequestAnalysis }) => {
   if (state.kind === "not_run") {
-    const button = (
+    return (
       <Button
         type="button"
         variant="secondary"
         size="sm"
-        isDisabled={!requestAnalysisEligible}
         onClick={() => onRequestAnalysis(vulnerabilityIdentifier)}
-        aria-label={
-          requestAnalysisEligible
-            ? `Request exploit intelligence analysis for ${vulnerabilityIdentifier}`
-            : `Request exploit intelligence analysis for ${vulnerabilityIdentifier} (available only for Critical and High severity)`
-        }
+        aria-label={`Request exploit intelligence analysis for ${vulnerabilityIdentifier}`}
       >
         Request Analysis
       </Button>
     );
-
-    if (!requestAnalysisEligible) {
-      return (
-        <Tooltip content={requestIneligibleTooltip}>
-          <span style={{ display: "inline-block" }}>{button}</span>
-        </Tooltip>
-      );
-    }
-
-    return button;
   }
 
   return (
