--- client/src/app/pages/sbom-details/components/exploit-intelligence.css
@@ -1,3 +1,0 @@
-.exploit-intelligence-in-progress-label .pf-v6-c-label__icon .pf-v6-c-spinner {
-  --pf-v6-c-spinner--Color: currentColor;
-}
