+++ client/src/app/queries/exploit-intelligence.ts
@@ -1,18 +1,10 @@
 import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
-import axios from "axios";
-import type { AxiosError } from "axios";
 
-import type {
-  ExploitIntelligenceCellState,
-  ExploitIntelligenceFinding,
-  ExploitIntelligenceFindingBreakdown,
-} from "@app/pages/sbom-details/components/exploit-intelligence-analysis-cell";
+import { client } from "@app/axios-config/apiInit";
 
 const EI_API = "/api/v3/exploit-intelligence";
 
 export const ExploitIntelligenceQueryKey = "exploit-intelligence";
-
-// -- API types ----------------------------------------------------------------
 
 interface SubmitAnalysisRequest {
   sbom_id: string;
@@ -50,116 +42,22 @@ interface JobsListResponse {
   total: number;
 }
 
-// -- State mapping ------------------------------------------------------------
-
-const buildBreakdown = (
-  job: ExploitIntelligenceJobSummary,
-): ExploitIntelligenceFindingBreakdown | undefined => {
-  if (job.product_id == null || job.total_components == null) {
-    return undefined;
-  }
-  return {
-    vulnerableCount: job.vulnerable_components ?? 0,
-    uncertainCount: job.uncertain_components ?? 0,
-    notVulnerableCount: job.not_vulnerable_components ?? 0,
-    failedCount: job.failed_components ?? 0,
-  };
-};
-
-const buildCompletedFinding = (
-  variant: Finding,
-  breakdown: ExploitIntelligenceFindingBreakdown | undefined,
-): ExploitIntelligenceFinding => {
-  switch (variant) {
-    case "vulnerable":
-      return {
-        variant: "vulnerable",
-        count: breakdown ? breakdown.vulnerableCount : undefined,
-        breakdown,
-      };
-    case "uncertain":
-      return {
-        variant: "uncertain",
-        count: breakdown ? breakdown.uncertainCount : undefined,
-        breakdown,
-      };
-    case "not_vulnerable":
-      return { variant: "not_vulnerable", breakdown };
-  }
-};
-
-export const jobToState = (
-  job: ExploitIntelligenceJobSummary,
-): ExploitIntelligenceCellState => {
-  if (job.status === "pending" || job.status === "running") {
-    return {
-      kind: "finding",
-      finding: { variant: "in_progress" },
-    };
-  }
-
-  if (job.status === "failed") {
-    return {
-      kind: "finding",
-      finding: { variant: "failed" },
-    };
-  }
-
-  // completed
-  const variant = job.finding ?? "uncertain";
-  const breakdown = buildBreakdown(job);
-  return {
-    kind: "finding",
-    finding: buildCompletedFinding(variant, breakdown),
-    reportUrl: job.report_url ?? undefined,
-  };
-};
-
-export const buildStateMap = (
-  jobs: ExploitIntelligenceJobSummary[],
-): Record<string, ExploitIntelligenceCellState> => {
-  const map: Record<string, ExploitIntelligenceCellState> = {};
-  for (const job of jobs) {
-    const existing = map[job.vulnerability_id];
-    if (!existing || shouldReplace(existing, job)) {
-      map[job.vulnerability_id] = jobToState(job);
-    }
-  }
-  return map;
-};
-
-const shouldReplace = (
-  current: ExploitIntelligenceCellState,
-  candidate: ExploitIntelligenceJobSummary,
-): boolean => {
-  if (current.kind === "not_run") return true;
-  if (candidate.status === "pending" || candidate.status === "running")
-    return true;
-  if (
-    candidate.status === "completed" &&
-    current.finding.variant !== "vulnerable" &&
-    current.finding.variant !== "not_vulnerable" &&
-    current.finding.variant !== "uncertain"
-  )
-    return true;
-  return false;
-};
-
 const POLLING_INTERVAL = 10_000;
 
 const hasActiveJobs = (jobs: ExploitIntelligenceJobSummary[]): boolean =>
   jobs.some((j) => j.status === "pending" || j.status === "running");
 
-// -- Hooks --------------------------------------------------------------------
-
 export const useFetchExploitIntelligenceJobs = (sbomId: string | undefined) => {
   const { data, isLoading, error } = useQuery({
     queryKey: [ExploitIntelligenceQueryKey, "jobs", sbomId],
     queryFn: () =>
-      axios.get<JobsListResponse>(`${EI_API}/jobs`, {
-        params: { sbom_id: sbomId, limit: 1000 },
+      client.get<JobsListResponse>({
+        url: `${EI_API}/jobs`,
+        query: { sbom_id: sbomId, limit: 1000 },
       }),
     enabled: !!sbomId,
+    refetchOnWindowFocus: true,
+    staleTime: 0,
     refetchInterval: (query) => {
       const items = query.state.data?.data?.items;
       if (items && hasActiveJobs(items)) {
@@ -173,10 +71,8 @@ export const useFetchExploitIntelligenceJobs = (sbomId: string | undefined) => {
 
   return {
     jobs,
-    stateMap: buildStateMap(jobs),
     isFetching: isLoading,
-    fetchError: error as AxiosError | null,
-    hasActiveJobs: hasActiveJobs(jobs),
+    fetchError: error,
   };
 };
 
@@ -185,7 +81,10 @@ export const useSubmitExploitAnalysisMutation = () => {
 
   return useMutation({
     mutationFn: (params: SubmitAnalysisRequest) =>
-      axios.post<SubmitAnalysisResponse>(`${EI_API}/analyze`, params),
+      client.post<SubmitAnalysisResponse>({
+        url: `${EI_API}/analyze`,
+        body: params,
+      }),
     onSuccess: async (_data, variables) => {
       await queryClient.invalidateQueries({
         queryKey: [ExploitIntelligenceQueryKey, "jobs", variables.sbom_id],
